formatted weather
